requirements.txt
# IndiaQuant MCP dependencies
mcp>=1.2.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
aiohttp>=3.9.0
uvicorn>=0.27.0
starlette>=0.36.0
python-dotenv>=1.0.0
pytz>=2023.3

# testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
